==> 2_summaryResultFigures.Rmd <==
---
title: "2_resultsPlots"
output: html_document
date: "2023-03-13"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Libraries
```{r}
library(tidyverse)
library(sf)
library(lubridate)
library(grDevices)
library(mapview)
library(extrafont)
library(ggpubr)
library(ggmap)
library(RgoogleMaps)
library(broom)
library(feather)
library(tidyhydat)
library(sp)
library(data.table)
library(ggalluvial)
library(patchwork)
library(magick)
library(units)
library(Kendall)
library(ggspatial)
library(dtplyr)
#Import libraries for Random Forest
library(caret) 
library(e1071)
library(Boruta)
library(tidymodels)
library(skimr)
library(vip)
```

# Set filenames & set constants
```{r}
# dates for version control
todayDate  = "20230324" # the first data join phase

# Set folder for all input data
int.wd = "~/WRR Submission 2 Data/Script 2"

#Name of file for lake shapefiles & island polygon shapefiles & river reach shapefiles
lakes.shapeFile = "mackenzieGoodLakes.shp"
import.sword = "na_sword_reaches_hb82_v14.shp"
setwd(int.wd)
lakes.sf = st_read(lakes.shapeFile)
# Piliouras delta islands shapefile
pil.file = "MackenzieDeltaIslands.shp"

# Image export folder
images.wd = "~/images"

```

# Import river centerlines and set the projection for all future plots, import classifications
```{r}
crs.plot = "+proj=tcea +lon_0=-134.3847656 +datum=WGS84 +units=m +no_defs"
setwd(int.wd)
study.area.large=cbind.data.frame(lon=c(-136.80, -136.80, -133.47, -133.47), 
                 lat=c(67.25, 69.55, 69.55, 67.46)) %>% 
  st_as_sf(coords=c("lon", "lat")) %>% st_set_crs(4326) %>% st_bbox() %>% 
  st_as_sfc() %>% 
  st_transform(crs = crs.plot)
mack.basin.large = st_read(import.sword) %>% 
  st_transform(crs = crs.plot) %>% 
  st_intersection(study.area.large) %>% dplyr::filter(width>90)

# import classifications
setwd(int.wd)
all.classified.filter = read_feather(paste0("final.class_", todayDate, ".feather")) 

# import functions
# takes the mode of a vector
getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}

```

# Plot a summary of connectivity through time.
## Figure 4 and Supplemental Figure 6
```{r}
# Prep data for plotting
plot.prep = all.classified.filter %>% lazy_dt() %>% #convert to lazy DT for faster processing
  mutate(.pred_class=as.numeric(as.character(.pred_class)))  %>% #convert classification from factor to numeric
  group_by(OBJECTID, month) %>% 
  summarise(mean.score=mean(.pred_class, na.rm=T),
            med.score = median(.pred_class, na.rm=T),
            sd.score = sd(.pred_class),
            mod.score = getmode(.pred_class),
            count= n() ) %>% dplyr::filter(count>=10) %>% ungroup() %>% 
  as_tibble() %>% 
  left_join(lakes.sf %>% dplyr::select(-count), 
            by="OBJECTID") %>% st_as_sf() %>% 
  st_transform(crs = crs.plot) %>% 
  dplyr::filter(month %in% c(5, 6, 7, 8, 9))  %>% 
  select(OBJECTID, mean.score, sd.score,med.score, mod.score,geometry, month) %>% 
  gather(Key, Value, -OBJECTID, -month, -geometry) %>% 
  mutate(monthPlot = case_when(
    month==5 ~ factor("May", levels=c("May","June", "July", "August", "September")),
    month==6 ~ factor("June", levels=c("May","June", "July", "August", "September")),
    month==7 ~ factor("July", levels=c("May","June", "July", "August", "September")),
    month==8 ~ factor("August", levels=c("May","June", "July", "August", "September")),
    month==9 ~ factor("September", levels=c("May","June", "July", "August", "September"))
  ) )%>% filter(month!=5)

# prepare plot set up## Set factors for month facet labels
scale_params <- tibble::tibble(
   monthPlot = factor("June", levels = c("June", "July", "August", "September")),
   Key ="mean.score"
  ) 


plot.labs.sd = data.frame(month=c(6,7,8,9), 
                            tag = c("e.", "f.", "g.", "h."), 
                            lat = rep(69.5, 4), lon = rep(-136.5,4)) %>% 
  st_as_sf(coords=c("lon", "lat")) %>% 
  st_set_crs(4326) %>% st_transform(crs.plot) %>% 
  mutate(monthPlot = case_when(
    month==5 ~ factor("May", levels=c("May","June", "July", "August", "September")),
    month==6 ~ factor("June", levels=c("May","June", "July", "August", "September")),
    month==7 ~ factor("July", levels=c("May","June", "July", "August", "September")),
    month==8 ~ factor("August", levels=c("May","June", "July", "August", "September")),
    month==9 ~ factor("September", levels=c("May","June", "July", "August", "September"))
  ))

plot.labs.mean = data.frame(month=c(6,7,8,9), 
                            tag = c("a.", "b.", "c.", "d."), 
                            lat = rep(69.5, 4), lon = rep(-136.5,4)) %>% 
  st_as_sf(coords=c("lon", "lat")) %>% 
  st_set_crs(4326) %>% st_transform(crs.plot) %>% 
  mutate(monthPlot = case_when(
    month==5 ~ factor("May", levels=c("May","June", "July", "August", "September")),
    month==6 ~ factor("June", levels=c("May","June", "July", "August", "September")),
    month==7 ~ factor("July", levels=c("May","June", "July", "August", "September")),
    month==8 ~ factor("August", levels=c("May","June", "July", "August", "September")),
    month==9 ~ factor("September", levels=c("May","June", "July", "August", "September"))
  ))


# Plot standard deviation in connectivity
sd.plot.lake = plot.prep %>% filter(Key=="sd.score") %>% 
  ggplot()+ geom_sf(aes(fill=Value), color=NA)+
  facet_wrap(~monthPlot, nrow=1)+theme_void()+
  geom_sf_text(data=plot.labs.sd,aes(label=tag))+
  scale_fill_viridis_c(option="plasma", limits=c(0,1), oob=squish,
                       labels=c(0, 0.25, 0.5, 0.75,1))+
  theme(
        strip.text = element_blank(),
        legend.text=element_text(size=10), 
        legend.position="right",
        legend.title = element_text(size=10, face="bold"))+
  labs(fill="Std. Dev.\nConnectivity\nScore")

# Plot mean connectivity
mean.plot.lake = plot.prep %>% filter(Key=="mean.score") %>% 
  ggplot()+ geom_sf(aes(fill=Value), color=NA)+
  facet_wrap(~monthPlot, nrow=1)+theme_void()+
  annotation_scale(text_cex = 0.9, text_col="black", data=scale_params)+
  geom_sf_text(data=plot.labs.mean,aes(label=tag))+
  scale_fill_viridis_c( limits=c(0,2), oob=squish,
                       labels=c(0, 0.5, 1, 1.5,2))+
  theme(
        strip.text = element_text(size=12, face="bold"),
        legend.text=element_text(size=10), 
        legend.position="right",
        legend.title = element_text(size=10, face="bold"))+
  labs(fill="Mean\nConnectivity\nScore")
mean.plot.lake/sd.plot.lake

setwd(images.wd)
ggsave(paste0(todayDate, "lakeMonthlySummary.png"), width=8, height = 6, units="in") 


# Plot summary statistics of connectivity through time in islands
setwd(int.wd)
mack.islands = st_read(pil.file) %>%
  dplyr::select(fid, geometry) %>%
  st_transform(crs.plot)

islands.join = lakes.sf %>% st_transform(crs.plot) %>% st_join(mack.islands) %>%
  filter(!is.na(fid)) %>% select(OBJECTID, fid, geometry)

island.summary = all.classified.filter %>% 
  mutate(.pred_class=as.numeric(as.character(.pred_class))) %>% lazy_dt() %>% 
  left_join(islands.join, by="OBJECTID") %>% filter(!is.na(fid)) %>% 
  group_by(month, fid) %>% 
  summarise(mean.con = mean(.pred_class),
            sd.con = sd(.pred_class),
            count=n()) %>% filter(month!=5) %>% 
  mutate(monthPlot = case_when(
    month==5 ~ factor("May", levels=c("May","June", "July", "August", "September")),
    month==6 ~ factor("June", levels=c("May","June", "July", "August", "September")),
    month==7 ~ factor("July", levels=c("May","June", "July", "August", "September")),
    month==8 ~ factor("August", levels=c("May","June", "July", "August", "September")),
    month==9 ~ factor("September", levels=c("May","June", "July", "August", "September"))
  ) ) %>% 
  left_join(mack.islands, by="fid") %>% as_tibble() %>% st_as_sf()

scale_params <- tibble::tibble(
   monthPlot = factor("June", levels = c("June", "July", "August", "September")),
   month =6
  ) 



mean.plot.islands = ggplot(data = island.summary )+
  geom_sf(aes(fill=mean.con), color=NA)+
  scale_fill_viridis_c( limits=c(0,2), oob=squish,
                       labels=c(0, 0.5, 1, 1.5,2))+
  geom_sf_text(data=plot.labs.mean,aes(label=tag))+
  facet_wrap(~monthPlot, nrow=1)+
  annotation_scale(text_cex = 0.9, text_col="black", data=scale_params)+
  theme_void()+
  theme(axis.text = element_blank(),
        strip.text = element_text(size=12, face="bold"),
        legend.text=element_text(size=10), 
        legend.position="right",
        legend.title = element_text(size=10, face="bold"), 
      panel.border=element_blank())+
  labs(fill="Mean\nConnectivity\nScore")

sd.plot.islands = ggplot(data = island.summary)+
  geom_sf(aes(fill=sd.con), color=NA)+
  scale_fill_viridis_c(option="plasma", limits=c(0,1), oob=squish,
                       labels=c(0, 0.25, 0.5, 0.75,1))+
  facet_wrap(~monthPlot, nrow=1)+
  geom_sf_text(data=plot.labs.sd,aes(label=tag))+
  theme_void()+
  theme(axis.text = element_blank(),
        strip.text = element_blank(),
        legend.text=element_text(size=10), 
        legend.position="right",
        legend.title = element_text(size=10, face="bold"),
      panel.border = element_blank(),
      panel.background = element_blank(),
      plot.background = element_blank())+
  labs(fill="Std. Dev.\nConnectivity\nScore")

mean.plot.islands / sd.plot.islands
setwd(images.wd)
ggsave(paste0(todayDate, "islandMonthlySummary.png"), width=8, height = 6, units="in")  
```

# plot number of observations per month
## Figure 2
```{r}
# prep data for plotting
plot.uncert.prep = all.classified.filter %>% 
  mutate(.pred_class = as.numeric(as.character(.pred_class))) %>% 
  as_tibble() %>% 
  group_by(OBJECTID, month) %>%count() %>% ungroup() %>% 
  left_join(lakes.sf, by="OBJECTID") %>% st_as_sf() %>% 
  st_transform(crs = crs.plot) %>% 
  mutate(monthWord = 
           case_when(month==5~"May",
                     month==6 ~"June", 
                     month==7 ~"July", 
                     month==8~"August",
                     month==9~"September"))
# set up factors so months plot in order 
plot.uncert.prep$monthWord = factor(plot.uncert.prep$monthWord, levels=c("May","June", "July", "August", "September"))

# set those names as factors - this is used to get an annotation only in the first plot (May)
scale_params <- tibble::tibble(
  monthWord = factor("May", levels = c("May","June", "July", "August", "September"))
)

plot.labs.uncert = data.frame(month=c(5,6,7,8,9), 
                            tag = c("a.", "b.", "c.", "d.", "e."), 
                            lat = rep(69.5, 5), lon = rep(-136.5,5)) %>% 
  st_as_sf(coords=c("lon", "lat")) %>% 
  st_set_crs(4326) %>% st_transform(crs.plot) %>% 
  mutate(monthWord = case_when(
    month==5 ~ factor("May", levels=c("May","June", "July", "August", "September")),
    month==6 ~ factor("June", levels=c("May","June", "July", "August", "September")),
    month==7 ~ factor("July", levels=c("May","June", "July", "August", "September")),
    month==8 ~ factor("August", levels=c("May","June", "July", "August", "September")),
    month==9 ~ factor("September", levels=c("May","June", "July", "August", "September"))
  ))
# Plot!
ggplot(data=plot.uncert.prep)+
    geom_sf(aes(fill=n), color=NA)+
    theme_void()+
    annotation_scale(text_cex = 0.9,
                     data=scale_params)+
    geom_sf_text(data=plot.labs.uncert,aes(label=tag))+
    #geom_sf(data=mack.basin.large, color="grey65")+
    scale_fill_viridis_c(begin=0, end=1, direction=1, oob=squish,limits=c(0,125),
                         breaks = c(0, 25, 50, 75, 100, 125),
                         labels=c("0","25", "50","75", "100", ">125"))+
    theme(strip.text = element_text(size=14, face="bold"),
          legend.title = element_text(size=12, face="bold"),
          legend.text=element_text(size=12),
          legend.position = "right",
          legend.title.align = 1)+labs(fill="Total Number\nof Observations")+
  guides(fill=guide_colourbar(direction="vertical",  title.position = "top", title.hjust = 0.5))+facet_wrap(~monthWord, nrow=1)

setwd(images.wd)
ggsave(paste0(todayDate, "_MonthlyUncertMap.png"), height=3, width=8, units="in", bg="white")
```

# plot study area figure
## Figure 1 
```{r}
# Prep for the large map
n.am <- sfheaders::sf_multipolygon(obj = map_data("world"),
                             multipolygon_id ="group",
                             x="long", y="lat",
                             keep = T) %>% 
  st_set_crs(4326) %>% filter(region=="Canada" | region=="USA") %>% 
  st_transform("ESRI:102008") 

coord.prep = st_coordinates(n.am) %>% as_tibble()

min.x = -4593551
max.x = max(coord.prep$X)
min.y = 0
max.y = 4017149

point = st_centroid(study.area.large) %>% st_transform("ESRI:102008")

largescale.map = ggplot()+
  geom_sf(data=n.am %>% 
            st_crop(xmin = min.x, xmax =max.x , ymin = min.y, ymax = max.y), 
          fill="grey20", color="grey50", size=0.1)+theme_bw()+
  geom_sf(data=point, color="#0096FF", size=3)+
  theme(axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill="grey90"),
        plot.background = element_rect(fill="transparent", color=NA))+
  coord_sf(expand = FALSE)



# plot the zoom in map
setwd(int.wd)
require(rnaturalearth)
world_ne <- ne_countries(scale = "large", returnclass = "sf")
setwd(int.wd)
world_noaa = st_read("GSHHS_h_L1.shp")
mack.delta <- world_noaa %>% 
  st_transform(crs.plot) %>% st_crop(xmin = -150941.35, ymin = 7486525.66,
                                     xmax = 39377.58, ymax = 7744125.65 )


locations = hy_stations(station_number=c(
  "10LC002", "10LC012", "10LC013", "10MC003", "10MC011", "10MC023"
)) %>% st_as_sf(coords=c("LONGITUDE", "LATITUDE")) %>% 
  st_set_crs(4326) %>%  st_transform(crs.plot)

zoom.map = ggplot()+
 # coord_sf(crs = st_crs())  # force the ggplot2 map to be in 3857
  geom_sf(data=mack.delta, fill="darkseagreen", color="darkseagreen4")+
  geom_sf(data=lakes.sf %>% st_transform(crs.plot), fill="grey99", color=NA, inherit.aes = F)+
  
  geom_sf(data=mack.basin.large%>% st_transform(crs.plot) %>% 
            st_intersection(mack.delta %>% st_transform(crs.plot)),
          color="cadetblue")+theme_bw()+
  geom_sf(data=locations, shape=17, color="black", inherit.aes = F, size=2)+
  theme(axis.text.x=element_text(angle=45, hjust = 1, size=12),
        axis.text.y=element_text(size=12),
        panel.background = element_rect(fill="lightblue"),
        panel.grid = element_line(colour="#7393B3"),
        legend.position="none")+
  annotation_scale(text_col="white")+
  coord_sf(expand = FALSE)
zoom.map

library(cowplot)
ggdraw() +
  draw_plot(zoom.map) +
  draw_plot(largescale.map,
    height = 0.2,
    x = -0.11,
    y = 0.18
  )
setwd(images.wd) #todayDate
ggsave(paste0("20230423","_studyArea.pdf"), device="pdf", width=4, height=5, units="in")
```

# Summarize connected lake area during the month after peak discharge at Arctic Red River
## Figure 5
```{r}
# Get yearly timing of peak flow at Arctic Red River
complete.flows = hy_daily_flows(
  station_number=c("10LC014"),
  start_date = "1973-01-01") %>% 
  mutate(doy = yday(Date),
         month = month(Date),
         year = year(Date))

# uses https://agupubs-onlinelibrary-wiley-com.libproxy.lib.unc.edu/doi/full/10.1002/2012WR013198 to define freshet initiation
freshet.initiation = complete.flows %>% arrange(Date) %>% filter(year>=1984) %>% 
  mutate(lag.value = lag(Value, n=1),
         diff = Value-lag.value,
         three.pct = Value*0.03,
         thresh.tf = diff>=three.pct) %>% 
  filter(thresh.tf==T & doy>31) %>% 
  group_by(year) %>%
  mutate(rnum= row_number()) %>% filter(rnum==1) %>% ungroup() %>% select(year, doy) %>% 
  rename(freshet.in = doy)

first.peak = complete.flows %>% filter(STATION_NUMBER=="10LC014") %>% 
  arrange(Date) %>% 
  left_join(freshet.initiation %>% select(year, freshet.in), by="year") %>% 
  filter(doy>freshet.in) %>% 
  mutate(lag.value = lag(Value, n=1),
         diff = Value-lag.value,
         three.pct = Value*0.03,
         thresh.tf = diff<= (-three.pct)) %>% 
  filter(diff<0 & Value>=10000) %>% 
  group_by(year) %>% 
  mutate(rnum= row_number()) %>% filter(rnum==1) %>% ungroup() %>% select(year, doy, Value) %>% 
  rename(first.peak = doy, peak.value=Value)

# filter lake classifications based on the month after the first discharge peak
filt.obs = all.classified.filter %>% 
  left_join(first.peak, by= "year") %>% filter(!is.na(first.peak)) %>% 
  filter(doy>=first.peak & doy<=first.peak+28)

filt.obs = all.classified.filter %>% select(OBJECTID, .pred_class, year, doy, month) %>% 
  left_join(first.peak, by= "year") %>% filter(!is.na(first.peak)) %>% 
  filter(doy>=first.peak)%>%
  mutate(timediff = doy-first.peak,
         timegroup = case_when(
           timediff <= 27  ~ 0,
           timediff >= 28 & timediff <= 55 ~ 28, 
           timediff >= 56 & timediff <= 83 ~ 56, 
           timediff >= 84 & timediff <= 111 ~ 84,
           timediff >= 112 & timediff <= 139 ~ 112)) %>% filter(!is.na(timegroup))


# Calculate lake areas
lakes.prep = lakes.sf %>% select(OBJECTID, geometry) %>% st_transform(crs.plot) 
lakes.prep$area = lakes.prep %>% st_area()
library(grid)
lakes.area = lakes.prep %>% 
  mutate(area.km = drop_units(area) * (1*10^-6)) 
total.area = sum(lakes.area$area.km)
total.area

# prep filtered obs for analysis 
getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}

# Summarize connected area by 2 week period after peak discharge
connected.area.prep = filt.obs %>% 
  select(.pred_class, OBJECTID, doy, month, year, timegroup) %>% 
  mutate(.pred_class=as.numeric(as.character(.pred_class))) %>% 
  group_by(OBJECTID, timegroup) %>% 
  summarise(mean.score=mean(.pred_class, na.rm=T),
            med.score = median(.pred_class, na.rm=T),
            mod.score = getmode(.pred_class),
            sd.score = sd(.pred_class),
            count= n() ) %>% dplyr::filter(count>=10) %>% 
  left_join(lakes.area, by="OBJECTID") %>% st_as_sf()
# Loop through using different thresholds to define connectivity
thresh = seq(0,2, by=0.25)
all.df = NULL
for (z in 1:length(thresh)){
  thrshhld = thresh[z]
  for (d in unique(connected.area.prep$timegroup)){
    con.area.df = connected.area.prep %>% 
    filter(mean.score>= thrshhld) %>% filter(timegroup==d)
    area.observed = connected.area.prep %>% 
      filter(timegroup==d)
    area.observed = sum(area.observed$area.km)
    num.lakes = nrow(con.area.df)
    con.area = sum(con.area.df$area.km)
    pct.area = con.area/area.observed
    con.vol = con.area* (1*0.001)
    con.vol.pct = (con.area* (1*0.001))/19.9
    combo.df = data.frame(thresh = thrshhld,con.area.km = con.area, 
                          pct.area = pct.area, con.vol.pct=con.vol.pct,
                          con.vol=con.vol, timegroup = d, 
                          observed.area = area.observed)
    all.df = rbind.data.frame(all.df, combo.df)
  }
}

# Plot connectivity pct area using a class 1 threshold
all.df = all.df %>% 
  mutate(timegroup2= case_when(
    timegroup==0 ~ "0 - 27",
    timegroup==28 ~ "28 - 55",
    timegroup==56 ~"56 - 83",
    timegroup==84 ~ "84 - 111",
    timegroup==112 ~ "112 - 139"
  )) 
all.df$timegroup2 = factor(all.df$timegroup2, levels = c("0 - 27", "28 - 55", "56 - 83", "84 - 111", "112 - 139"))

all.df%>% 
  ggplot(aes(x=thresh, y=pct.area, color = as.factor(timegroup2),
             group=timegroup2))+geom_line()+geom_point()+
  theme_bw()+ylab("probability of\nexceeding connectivity threshold")+
  xlab("mean connectivity class threshold")+
  theme(axis.text = element_text(size=12),
        axis.title = element_text(size=12, face="bold"),
        legend.title = element_text(size=12, face="bold"),
        legend.text=element_text(size=12))+
  scale_x_continuous(breaks=c(0,0.5, 1,1.5, 2))+
  scale_color_manual(values=c("#130112", "#A21675", "#E74BB5", "#F093D3", "#96788D"))+
  labs(color="Days after\npeak discharge")

setwd(images.wd)
ggsave(paste0(todayDate, "probabilityOfExceed.pdf"), width=6, height=4, units = "in")
# plot variability of flow - NOT IN PAPER
## calculate recurrance intervals
RI.df = complete.flows %>% 
  group_by(year) %>% 
  summarise(max.val = max(Value, na.rm=T)) %>% 
  arrange(desc(max.val)) %>% 
  mutate(rank = row_number(),
         RI = (47+1)/rank,
         diff1_5 = abs(1.5-RI)) %>% # difference from 1.5yr recurrance interval
  arrange(diff1_5)
RI.dis = RI.df[1,"max.val"][[1]] # get RI closes to 1.5 years (1.48)
## calculate mean annual flood volumes - i.e. volume above the 1.5yr RI flood
test = complete.flows %>% 
  filter(Value>=RI.dis) %>% 
  mutate(flood.vol.cms = Value,
         flood.vol.ckms = flood.vol.cms*1e-9, #m^3 to km^3
         flood.vol.ckmd = flood.vol.ckms * 86400# sec to day
    #daily.total.vol = flood.vol 
    ) %>% 
  group_by(year) %>% 
  summarise(flood.vol = sum(flood.vol.ckmd))

mean(test$flood.vol)
sd(test$flood.vol)

ggplot(data=test, aes(x=year, y=flood.vol))+
  geom_point()+geom_line()+
  theme_bw()+ylab(expression(annual~flood~volume~'('~km^3~')'))+
  theme(axis.text = element_text(size=12),
        axis.title = element_text(size=12, face="bold"))+
  xlim(1973, 2021)
setwd(images.wd)
ggsave(paste0(todayDate, "annualFloodVol.pdf"), width=4, height=4, units = "in")
```
